# controller/user_services/model/register.model.go
package usermodel

type Register struct {
	Name            string `db:"name" json:"name"`
	Email           string `db:"email" json:"email"`
	Password        string `db:"password" json:"password"`
	ConfirmPassword string `json:"confirm_password"`
}